old-tests/submission/Infraware/Forms/contents/Forms/label_htmlFor.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>label_htmlFor</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form"
		  name="input_form">
		<fieldset id="input_field">
			<legend id='input_legend'><label id='input_label'></label></legend>
 		</fieldset>
	</form>
	<script>

		var label = document.getElementById("input_label");

		if (typeof(label.htmlFor) == "string") {
			test(function() {
				assert_equals(label.htmlFor, "", "htmlFor attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("htmlFor attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
